requirements.txt
absl-py==2.2.0
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.3
bitsandbytes==0.45.3
bitsandbytes-cuda111==0.26.0.post2
bitsandbytes-windows==0.37.5
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
colorama==0.4.6
datasets==3.4.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
gpt4all==2.8.2
grpcio==1.71.0
gutenbergpy==0.3.5
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpsproxy_urllib2==1.0
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
keras==3.9.0
libclang==18.1.1
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.0
openai==1.68.2
openai-whisper==20231117
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
srt==3.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.50.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vosk==0.3.45
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==1.0.2
yt-dlp==2025.2.19
flask==2.3.2
flask-cors==3.0.10
whisper-openai==1.1.5
tensorflow==2.10.0
numpy==1.23.5
python-dotenv==0.21.0
waitress==2.1.2
